services/quiz/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.0
psycopg2-binary==2.9.9
openai==1.3.0
boto3==1.29.0
